requirements.txt
absl-py==0.8.0
apache-beam==2.15.0
astor==0.8.0
avro==1.9.0
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
crcmod==1.7
decorator==4.4.0
dill==0.2.9
docopt==0.6.2
enum34==1.1.6
fastavro==0.21.24
fasteners==0.15
funcsigs==1.0.2
future==0.17.1
futures==3.3.0
gast==0.2.2
geographiclib==1.49
geopy==1.20.0
google-api-core==1.14.2
google-apitools==0.5.28
google-auth==1.6.3
google-cloud-bigquery==1.17.0
google-cloud-bigtable==0.32.2
google-cloud-core==1.0.3
google-cloud-datastore==1.7.4
google-cloud-pubsub==0.39.1
google-pasta==0.1.7
google-resumable-media==0.3.3
googleapis-common-protos==1.6.0
googledatastore==7.0.2
grpc-google-iam-v1==0.11.4
grpcio==1.23.0
h5py==2.9.0
hdfs==2.5.8
httplib2==0.12.0
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
linecache2==1.0.0
Markdown==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
monotonic==1.5
numpy==1.16.5
oauth2client==3.0.0
pandas==0.24.2
pathlib2==2.3.4
pbr==5.4.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==1.0.16
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.9.1
ptyprocess==0.6.0
pyarrow==0.14.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pydot==1.4.1
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyVCF==0.6.8
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
scandir==1.10.0
scikit-learn==0.20.4
scipy==1.2.2
simplegeneric==0.8.1
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-data-validation==0.14.1
tensorflow-estimator==1.14.0
tensorflow-metadata==0.14.0
tensorflow-transform==0.14.0
termcolor==1.1.0
traceback2==1.4.0
traitlets==4.3.2
typing==3.6.6
unittest2==1.1.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
